Moving some code to GameViewModel class
