+++ app/build.gradle
@@ -58,4 +58,6 @@ dependencies {
     // Navigation
     implementation "android.arch.navigation:navigation-fragment-ktx:1.0.0-rc02"
     implementation "android.arch.navigation:navigation-ui-ktx:1.0.0-rc02"
+
+    implementation 'androidx.lifecycle:lifecycle-extensions:2.0.0'
 }
